[SPARK-41246][CORE] Widen RDDBlockId, internal maps, and event filters to Long RDD IDs
Extends the Phase-1 implementation to eliminate all Phase-1 compromises:

- RDDBlockId.rddId: Int -> Long; regex updated to (-?[0-9]+) to parse
  legacy negative IDs in old event logs
- All RDDBlockId construction sites use rdd.longId (RDD.scala,
  DAGScheduler.scala, LocalRDDCheckpointData.scala)
- Internal Int-keyed structures widened to Long (no MiMa impact):
  - AppStatusListener.liveRDDs: HashMap[Int] -> HashMap[Long]
  - ExecutorMonitor.cachedBlocks: HashMap[Int] -> HashMap[Long]
  - BasicEventFilterBuilder/SQLEventFilterBuilder.stageToRDDs inner
    Set[Int] -> Set[Long]; JobEventFilter/SQLLiveEntitiesEventFilter
    liveRDDs param widened to Set[Long]
- All <= Int.MaxValue truncation guards removed
- RDDInfo gains longId: Long and parentLongIds: Seq[Long] fields
- MemoryStore.getRddId: Option[Int] -> Option[Long]
- Redundant .toLong casts removed from BlockInfoManager,
  BlockManagerMasterEndpoint, StorageUtils
- MiMa exclusions for RDDBlockId added to v50excludes
- Config version corrected to 5.0.0

Test updates:
- LongRddIdSuite: filterWithLiveRDDs takes Set[Long]; true D-scenario
  StorageUtils test added; BasicEventFilter D test now asserts Long IDs
  are accepted
- BlockInfoManagerSuite D: now asserts blocks found (not empty)
- ExecutorMonitorSuite D: rddUpdate takes Long; asserts block removed
- DAGSchedulerSuite: cacheLocations key (Int,Int) -> (Long,Int)
- BasicEventFilterSuite, SQLLiveEntitiesEventFilterSuite: liveRDDs
  literal sets use Long

Feasibility experiment: full build passes on JDK 17/Scala 2.13 and
JDK 24/Scala 2.12.

+++ core/src/main/scala/org/apache/spark/storage/BlockInfoManager.scala
@@ -726,7 +726,7 @@ private[storage] class BlockInfoManager(trackingCacheVisibility: Boolean = false
 
     blockId match {
       case rddBlockId: RDDBlockId =>
-        doRemove(rddToBlockIds, rddBlockId.rddId.toLong, rddBlockId)
+        doRemove(rddToBlockIds, rddBlockId.rddId, rddBlockId)
       case broadcastBlockId: BroadcastBlockId =>
         doRemove(broadcastToBlockIds, broadcastBlockId.broadcastId, broadcastBlockId)
       case cacheId: CacheId =>
